backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
groq==0.4.0
pypdf==3.17.0
pdfplumber==0.11.0
requests==2.31.0
python-multipart==0.0.6
aiofiles==23.2.1
chromadb==0.4.24
numpy<2.0.0
sentence-transformers>=3.0.0
aiohttp>=3.9.0
